pyproject.toml
[project]
name = "bigquery-mcp"
version = "0.0.1"
description = "An MCP server for BigQuery"
authors = [
    {name="Padraig Alton", email="padraig.alton@justpark.com"},
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "db-dtypes>=1.4.2",
    "google-cloud-bigquery>=3.31.0",
    "google-cloud-bigquery-storage>=2.30.0",
    "mcp[cli]>=1.3.0",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
bigquery-mcp = "src.server:server.run"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pandas-stubs>=2.2.3.250308",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-watcher>=0.4.3",
    "ruff>=0.9.10",
    "sqlfluff>=3.3.1",
]
explore = [
    "jupyter>=1.1.1",
]

[tool.uv]
default-groups = [
    "dev",
    "explore",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.mypy]
warn_unused_configs = true
overrides = [
    { module = "saneyaml.*", ignore_missing_imports = true },
]

[tool.ruff]
target-version = "py313"
line-length = 89

[tool.ruff.lint]
select = [
    "ALL",
]
ignore = [
    "COM812",  # Trailing comma check not recommended (we use the ruff formatter)
    "D203",    # We don't like blank lines before class docstrings (incompatible with D211)
    "D205",    # Allow function summaries of more than one line
    "D212",    # We prefer (an aesthetic/readability choice) function summaries below triple quotes, not alongside
    "D400",    # We accept missing trailing full-stops at the end of one-liner function summaries
    "D415",    # We accept missing trailing punctuation at the end of one-liner function summaries
    "FIX002",  # We don't flag TO-DOs with ruff
    "RET504",  # We often want an 'unnecessary' assignment, as a return variable having a name can be clearer
    "SIM103",  # We accept 'unnecessary' boolean checks on the grounds that they can improve readability
]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "ANN001",   # Tests do not need to be annotated
    "ANN201",   # Tests do not need to be annotated
    "D",        # Ignore all documentation-related rules
    "SLF001",   # Ignore private member access
    "PLR2004",  # Ignore magic numbers in tests
    "S101",     # Ignore assert statements
]

[tool.ruff.lint.flake8-annotations]
suppress-none-returning = true

[tool.ruff.lint.isort]
lines-between-types = 1

[tool.sqlfluff.core]
dialect = "bigquery"
templater = "python"
exclude_rules = [
    "LT08"  # We don't want blank lines between CTEs
]

[tool.sqlfluff.rules.aliasing.table]
table = "explicit"

[tool.sqlfluff.rules.aliasing.column]
column = "explicit"

[tool.sqlfluff.rules.capitalisation.keywords]
capitalisation_policy = "upper"

[tool.sqlfluff.rules.capitalisation.identifiers]
capitalisation_policy = "lower"

[tool.sqlfluff.rules.capitalisation.functions]
extended_capitalisation_policy = "upper"

[tool.sqlfluff.rules.capitalisation.literals]
capitalisation_policy = "lower"

[tool.sqlfluff.rules.capitalisation.types]
extended_capitalisation_policy = "upper"

[tool.sqlfluff.rules.convention.select_trailing_comma]
select_clause_trailing_comma = "require"

[tool.sqlfluff.templater.python.context]
region = "europe-west2"  # template stub value for testing
